requirements.txt
pandas==1.5.3
torch==2.6.0
transformers==4.31.0
bitsandbytes==0.37.0
huggingface_hub==0.16.4
googletrans==4.0.0-rc1
gTTS==2.2.3
IPython==7.34.0
grpcio==1.59.0
grpcio-tools==1.59.0
streamlit==1.29.0
numpy>=1.20.0
protobuf==4.21.12
